LayoutTests/fast/loader/redirect-to-invalid-url-using-javascript-disallowed-expected.txt

main frame - didFinishDocumentLoadForFrame
main frame - didHandleOnloadEventsForFrame
main frame - didFinishLoadForFrame
Tests that we do not redirect to an invalid URL initiated by JavaScript. This test PASSED if you see an entry in the dumped frame load callbacks of the form: "willPerformClientRedirectToURL: http://A=a%B=b" followed by "didCancelClientRedirectForFrame".

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS caughtException is true
PASS successfullyParsed is true

TEST COMPLETE
